requirements.txt
# CodeJanitor 2.0 - Best-in-Class Tech Stack

# Orchestration & Agents
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0

# Configuration Management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Container Execution
docker>=7.0.0

# Git Operations
gitpython>=3.1.0
pygithub>=2.1.0

# Resilience & Retries
tenacity>=8.2.0

# CLI Interface
rich>=13.0.0

# Code Parsing
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0

# Graph & Knowledge Base
networkx>=3.0

# Vector Store (for future RAG)
chromadb>=0.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
